Add files via upload
Python graphs that compares Stephen Curry's minutes played in games and the final score of both the Golden State Warriors, or the opposing team. Curry is known as one of the most dominant point guards to ever play in the NBA. I was surprised to see that over the course of his career, with all the minutes he plays in games, the Warriors legend is very consistent at keeping the games close.